core/qareen/requirements.txt
# Qareen — single process requirements
# Install: pip install -r requirements.txt

# Core
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
pydantic>=2.0
pyyaml>=6.0
setproctitle>=1.3

# Audio (voice pipeline)
sounddevice>=0.5
numpy>=2.0
onnxruntime>=1.19
mlx-whisper>=0.4

# Database
# (stdlib sqlite3 is sufficient — no extra deps)

# Search
# (QMD is called via subprocess — no Python dep)
